move: 파일 위치 변경 config > exception
