+++ src/main/java/com/sparta/outsourcing/exception/ApplicationException.java
@@ -1,4 +1,4 @@
-package com.sparta.outsourcing.config;
+package com.sparta.outsourcing.exception;
 
 public class ApplicationException extends RuntimeException {
